requirements.txt
flask
flask-cors
google-genai
pydantic
